# api/openapi.yaml
openapi: 3.0.3
info:
  title: GreenPass 软件测试治理平台 API
  description: |
    测试质量管控平台后端契约（P1）。
    租户隔离：所有业务请求必须携带 `x-gp-team-id` 与可选 `x-gp-user-id` 头，经 RLS 过滤。
    审计哈希链 / 成本明细 / 门禁判定由 trusted 域统一写（口径单点）。
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /targets:
    get:
      summary: 被测对象树（工程/服务组/服务/模块）
      operationId: listTargets
      parameters:
        - { name: x-gp-team-id, in: header, required: true, schema: { type: integer } }
      responses:
        '200': { description: 树形列表, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Target' } } } } }
    post:
      summary: 创建被测对象节点
      operationId: createTarget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name, type ]
              properties:
                name: { type: string }
                type: { type: string, enum: [ project, service_group, service, module ] }
                parent_id: { type: integer }
                kind: { type: string }
      responses:
        '201': { description: 已创建, content: { application/json: { schema: { $ref: '#/components/schemas/Target' } } } }
  /targets/{id}/repo:
    post:
      summary: 关联上游仓库（来源溯源）
      operationId: attachRepo
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ url ]
              properties:
                url: { type: string, description: 仓库地址（如 git@github.com:org/repo.git） }
                default_branch: { type: string }
      responses:
        '200': { description: 已关联, content: { application/json: { schema: { $ref: '#/components/schemas/StatusResponse' } } } }
  # ---------- 用例（governance，GP1-02） ----------
...
  /gates/rules:
    put:
      summary: 装载门禁策略（版本化，策略即代码）
      operationId: upsertGateRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ target_id, scenario_id, rego ]
              properties:
                target_id: { type: integer }
                scenario_id: { type: integer }
                rego: { type: string, description: 声明式策略 JSON 或 Rego }
      responses:
        '200': { description: 已装载（含新版本号）, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
  /gates/evaluate:
    post:
      summary: 门禁判定（pass/fail/blocked，入审计链）
      operationId: evaluateGate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ run_id ]
              properties:
                run_id: { type: integer }
      responses:
        '200':
          description: 判定结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { type: string, enum: [ pass, fail, blocked ] }
  /gates/results:
    get:
      summary: 运行的门禁判定结果
      operationId: gateResults
      parameters:
        - { name: run_id, in: query, required: true, schema: { type: integer } }
      responses:
        '200': { description: 判定列表, content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } }
  # ---------- 成本（trusted，GP1-06） ----------
  /cost/record:
    post:
      summary: 成本计量写入（幂等：request_id+biz_point 不双计）
      operationId: recordCost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ request_id, biz_point ]
              properties:
                request_id: { type: string }
                biz_point: { type: string, enum: [ case_generate, case_execute, gate_judge, analysis ] }
                target_id: { type: integer }
                run_id: { type: integer }
                case_id: { type: integer }
                model: { type: string }
                tokens_in: { type: integer }
                tokens_out: { type: integer }
                unit_price: { type: number, description: 每千 token 单价 }
      responses:
        '200':
          description: inserted 是否新落账（false=幂等跳过）
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted: { type: boolean }
...
  /cost/items:
    get:
      summary: 成本明细（按大类/细类筛选）
      operationId: costItems
      parameters:
        - { name: category, in: query, schema: { type: string, enum: [ generate, execute ] } }
        - { name: point, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200': { description: 明细列表, content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } }
  /cost/compare/{caseId}:
    get:
      summary: 单用例成本历史对比（存量持平/递减证明）
      operationId: costCompare
      parameters:
        - { name: caseId, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 该用例各次成本时序, content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } }
  # ---------- 报告（trusted，GP1-07） ----------
  /reports/generate:
    post:
      summary: 由运行生成测试报告
      operationId: generateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ run_id ]
              properties:
                run_id: { type: integer }
                kind: { type: string, enum: [ project, service, case ], description: 默认 service；跨场景合编 P2 }
      responses:
        '200': { description: 已生成报告（含 HTML）, content: { application/json: { schema: { $ref: '#/components/schemas/Report' } } } }
  /reports/bundle:
    post:
      summary: 跨场景合编报告（HTML/Markdown）
      operationId: bundleReports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ report_ids ]
              properties:
                report_ids: { type: array, items: { type: integer } }
                format: { type: string, enum: [ html, markdown ], default: html }
      responses:
        '200': { description: 合编报告内容, content: { application/json: { schema: { $ref: '#/components/schemas/ReportBundle' } } } }
  /reports/{id}/export:
    get:
      summary: 导出报告
      operationId: exportReport
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: fmt, in: query, schema: { type: string, enum: [ html ], default: html } }
      responses:
        '200':
          description: HTML 报告
          content:
            text/html: {}
components:
  schemas:
    StatusResponse:
      type: object
      required: [ status ]
      properties:
        status: { type: string }
    Target:
      type: object
      required: [ id, team_id, type, name, kind, status ]
      properties:
        id: { type: integer, format: int64 }
        team_id: { type: integer, format: int64 }
        type: { type: string }
        name: { type: string }
        kind: { type: string }
        status: { type: string }
        parent_id: { type: integer, format: int64, nullable: true }
        repo_id: { type: integer, format: int64, nullable: true }
...
    Report:
      type: object
      required: [ id, team_id, run_id, target_id, kind, title, version, branch, scenario, status, summary, html, created_at ]
      properties:
        id: { type: integer, format: int64 }
        team_id: { type: integer, format: int64 }
        run_id: { type: integer, format: int64 }
        target_id: { type: integer, format: int64 }
        kind: { type: string }
        title: { type: string }
        version: { type: string }
        branch: { type: string }
        scenario: { type: string }
        status: { type: string }
        summary: { type: object, additionalProperties: true }
        evidence: { type: object, additionalProperties: true }
        html: { type: string }
        created_at: { type: string, format: date-time }
    ReportBundle:
      type: object
      required: [ reports, total, pass, fail ]
      properties:
        reports: { type: array, items: { $ref: '#/components/schemas/Report' } }
        total: { type: integer }
        pass: { type: integer }
        fail: { type: integer }
        html: { type: string }
        markdown: { type: string }
  securitySchemes:
    teamHeader:
      type: apiKey
      in: header
      name: x-gp-team-id
security:
  - teamHeader: []